+++ src/models/Settings.php
@@ -125,6 +125,13 @@ class Settings extends Model
     public bool $widgetEnabledCraftNewEntry = false;
 
     /**
+     * Enables the Admin Bar Widget: Authors
+     *
+     * @var bool
+     */
+    public bool $widgetEnabledCraftAuthors = false;
+
+    /**
      * Enables the Admin Bar Widget: New Entry
      *
      * @var bool