+++ CHANGELOG.md
@@ -4,9 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 5.2.0 - 2024-12-23
+- Added new Admin Bar Widget: Authors
+  - Lets you quickly see who the author was for the current page you are on—without needing to go to the edit page.
+
+### Changed
+- Refactored `AdminBarWidget` class for simpler maintainence and to remove unneeded queries.
+- Updated Admin Bar Component to version `1.1.4`
+
+
 ## 5.1.1 - 2024-12-20
 ### Fixed
 - Fixed issue with Blitz widget not showing cache status.
+
 
 ## 5.1.0 - 2024-12-20
 ### Added